pyproject.toml
[tool.poetry]
name = "community"
version = "0.1.0"
description = ""
authors = ["Michel NAUD <michel.naud26400@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic-settings = "^2.6.0"
python-frontmatter = "^1.1.0"
pyyaml = "^6.0.2"
jinja2 = "^3.1.0"
markdown = "^3.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
pytest-cov = "^6.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
